requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
neo4j==5.12.0
numpy==1.25.2
pandas==2.1.0
psycopg2==2.9.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.7
